pyproject.toml
[project]
name = "github-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.8",
    "dotenv>=0.9.9",
    "fastapi>=0.115.13",
    "openai>=1.90.0",
    "pydantic>=2.11.7",
    "pydantic-ai-slim[mcp]>=0.3.1",
    "pydantic-ai[mcp]>=0.3.2",
    "requests>=2.32.4",
    "uvicorn>=0.34.3",
]
